requirements.txt
outline-vpn-api
python-decouple
pyinstaller
openpyxl
requests
